Feature alert

+++ aiot-vision-collector/src/main/java/com/sandy/aiot/vision/collector/vo/TimeSeriesDataModelRsp.java
@@ -20,7 +20,7 @@ public class TimeSeriesDataModelRsp {
     @AllArgsConstructor
     public static class PredictionPointRsp {
         private String timestamp;
-        private Float value;
+        private Double value;
     }
 
     public static TimeSeriesDataModelRsp empty(){
